Move resolve(...) called in all add(...) to CatalogValidationRules
Use CatalogValidationRules to resolve initial valid state of CatalogInfo
objects being added to the catalog, and make good use of that `isNew`
flag, instead of the logic being scattered all over the place.
